Use common parser for asap and ase
